Move MDX Vite plugin to new file
